requirements.txt
streamlit
boto3
numpy
faiss-cpu
python-docx
PyPDF2
langchain_community
langchain
langchain-aws
awscli
websockets